cookiecutter/{{cookiecutter.module_name}}/requirements/dev.txt
bandit==1.6.2
black==19.3b0
Faker==1.0.4
factory-boy==2.11.1
flake8==3.7.7
isort==4.3.16
pylint==2.3.1
pylint-django==2.0.6
pytest==4.3.1
pytest-cov==2.6.1
pytest-django==3.4.8
